refactor(referentiels): un dossier par endpoint de documents de labellisation
`list-documents-labellisation/` hebergeait deux procedures la ou la convention
du depot veut un dossier par feature. Le nom du dossier ne couvrait que 5 de ses
9 fichiers, et aucune des deux procedures ne s'appelait comme lui.

`list-documents-audit/` et `list-documents-demande-labellisation/` sont
desormais symetriques de leurs deux voisins : un prefixe unique, egal au nom du
dossier. Les deux endpoints ne partageaient aucune logique — chacun resout son
objet, controle l'acces, appelle son repository et parse sa sortie.

Le montage du cycle de test, lui, servait les deux : il devient
`documents-labellisation.test-fixture.ts` plutot que d'etre duplique. Les six
tests e2e se repartissent trois et trois.

Les deux routeurs declarent leur gestionnaire d'erreur en champ de classe, comme
leurs voisins, au lieu de le reconstruire a chaque requete.
