+++ apps/backend/src/referentiels/referentiels.router.ts
@@ -13,7 +13,8 @@ import { HistoriqueRouter } from './historique/historique.router';
 import { CreatePreuveRouter } from './labellisations/create-preuve/create-preuve.router';
 import { GetLabellisationRouter } from './labellisations/get-labellisation.router';
 import { HandleMesureAuditStatutRouter } from './labellisations/handle-mesure-audit-statut/handle-mesure-audit-statut.router';
-import { ListDocumentsLabellisationRouter } from './documents/list-documents-labellisation/list-documents-labellisation.router';
+import { ListDocumentsAuditRouter } from './documents/list-documents-audit/list-documents-audit.router';
+import { ListDocumentsDemandeLabellisationRouter } from './documents/list-documents-demande-labellisation/list-documents-demande-labellisation.router';
 import { RequestLabellisationRouter } from './labellisations/request-labellisation/request-labellisation.router';
 import { StartAuditRouter } from './labellisations/start-audit/start-audit.router';
 import { UpdateAuditReportRouter } from './labellisations/update-audit-report/update-audit-report.router';
@@ -44,7 +45,8 @@ export class ReferentielsRouter {
     private readonly requestLabellisation: RequestLabellisationRouter,
     private readonly createPreuve: CreatePreuveRouter,
     private readonly validateAudit: ValidateAuditRouter,
-    private readonly listDocumentsLabellisationRouter: ListDocumentsLabellisationRouter,
+    private readonly listDocumentsAuditRouter: ListDocumentsAuditRouter,
+    private readonly listDocumentsDemandeLabellisationRouter: ListDocumentsDemandeLabellisationRouter,
     private readonly listDocumentsReferentielRouter: ListDocumentsReferentielRouter,
     private readonly listDocumentsMesureRouter: ListDocumentsMesureRouter,
     private readonly updateAuditReport: UpdateAuditReportRouter,
@@ -98,7 +100,8 @@ export class ReferentielsRouter {
 
     documents: this.trpc.mergeRouters(
       this.listDocumentsReferentielRouter.router,
-      this.listDocumentsLabellisationRouter.router,
+      this.listDocumentsAuditRouter.router,
+      this.listDocumentsDemandeLabellisationRouter.router,
       this.listDocumentsMesureRouter.router
     ),
